carrito-flow/carrito/requirements.txt
Flask==1.0.2
simplejson==3.13.2
google-api-python-client==1.7.8
google-cloud-pubsub==0.39.1
